Create README.md
This app has been developed java and it is an e-commerce app and specially developed for seller. With this app, one can upload product image to server . There is another app that has been developed for for buyer is complementary to each other. The uploaded  product images can be seen and be ordered from that app.

There is a special feature that is real time ordering system and PUSHER has been used as  the real time communication server.